+++ lib/src/client_stream.dart
@@ -110,6 +110,7 @@ class TusStreamClient extends TusBaseClient {
     super.url,
     super.uploadUrl,
     super.chunkSize,
+    super.progressSliceSize,
     super.tusVersion,
     super.cache,
     super.headers,